RPMB: fix read_cid() and data access with a real eMMC device
Prior to this commit, the RPMB code was tested in emulation mode only
(RPMB_EMU=1). This fixes a couple of issues found when testing with a
real eMMC device on a HiKey board. The eMMC is a Hardkernel 8GB module
plugged into the microSD port. Kernel is from Linaro's 96boards
repository [2], slightly modified so that the eMMC RPMB can be used on
the microSD connector (MMC_CAP_CMD23 is added to the capability flags
of the second MMC controller).

[2] https://github.com/96boards/linux branch hikey commit 54e6524885a9
("Merge pull request #178 from xin3liang/hikey-3.18-720p-boot-issue").

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>