tee-supplicant: rpmb: introduce readn() wrapper to the read() syscall
read_cid() obtains the ID of the eMMC device by reading from sysfs with
the read() function. Two bytes are read at a time but short reads (i.e.,
when only one byte is returned) and EINTR are not handled. Although I
*think* these situation cannot happen with sysfs, I was unable to find
any guarantee that it is the case. Therefore, introduce a readn()
function which does exactly that.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>